+++ packages/dnd/src/QueryBuilderDnD.tsx
@@ -78,7 +78,7 @@ export const QueryBuilderDnD = (props: QueryBuilderDndProps): React.JSX.Element 
           canDrop={props.canDrop}
           copyModeModifierKey={props.copyModeModifierKey}
           groupModeModifierKey={props.groupModeModifierKey}
-          noDragPreview={props.noDragPreview}>
+          hideDefaultDragPreview={props.hideDefaultDragPreview}>
           {props.children}
         </QueryBuilderDndWithoutProvider>
       </QueryBuilderContext.Provider>
@@ -108,7 +108,11 @@ export const QueryBuilderDndWithoutProvider = (props: QueryBuilderDndProps): Rea
   );
   const enableDragAndDrop = preferProp(true, props.enableDragAndDrop, rqbContext.enableDragAndDrop);
   const debugMode = preferProp(false, props.debugMode, rqbContext.debugMode);
-  const noDragPreview = preferProp(false, props.noDragPreview, rqbDndContext.noDragPreview);
+  const hideDefaultDragPreview = preferProp(
+    false,
+    props.hideDefaultDragPreview,
+    rqbDndContext.hideDefaultDragPreview
+  );
   const canDrop = preferAnyProp(undefined, props.canDrop, rqbDndContext.canDrop);
   const key = enableDragAndDrop && dnd ? 'dnd' : 'no-dnd';
 
@@ -163,7 +167,7 @@ export const QueryBuilderDndWithoutProvider = (props: QueryBuilderDndProps): Rea
       canDrop,
       copyModeModifierKey,
       groupModeModifierKey,
-      noDragPreview,
+      hideDefaultDragPreview,
       useDrag,
       useDrop,
     }),
@@ -172,7 +176,7 @@ export const QueryBuilderDndWithoutProvider = (props: QueryBuilderDndProps): Rea
       canDrop,
       copyModeModifierKey,
       groupModeModifierKey,
-      noDragPreview,
+      hideDefaultDragPreview,
       useDrag,
       useDrop,
     ]